Updating code from course

+++ src/serialize/Location.java
@@ -4,19 +4,19 @@ public class Location {
     private double lat;
     private double lng;
 
-    public int getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public int getLng() {
+    public double getLng() {
         return lng;
     }
 
-    public void setLat(int lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public void setLng(int lng) {
+    public void setLng(double lng) {
         this.lng = lng;
     }
 }
